dom/media/webvtt/test/mochitest/test_bug957847.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=957847
-->
<head>
  <meta charset='utf-8'>
  <title>Regression test for bug 957847 - Crash on TextTrack::AddCue </title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var trackElement = document.createElement('track');
trackElement.track.addCue(new VTTCue(0, 1, "A"));

// We need to assert something for Mochitest to be happy.
ok(true);
SimpleTest.finish();
</script>
</pre>
</body>
</html>
